Fix missing definitions
If a definition was not directly referenced as a variable type (e.g., it
was used via a $ref), then the definition was not pulled into the
schema.

Now we just pull all found types into the schema. This can bloat the
schema with unused definitions, but it is a safe option.