Template: Support deep merge of JSON templates
A new contentType has been added to denote the MIME type of the rendered
output. This allows us to select a JSON merge strategy if the
contentType is 'application/json'.

This merge is not currently configurable. We will also need to expose a
way for an end user to add new merge strategies.