template: Add support for calculating parameter values from arithmetic expressions
This is done with a new `mathExpression` property available when
defining parameters. Other parameters can be referenced in the
expression and will get included in the schema as if they appeared in
the template text. By default, any parameter defined with a
`mathExpression` property and no `format` will be given a `format` of
`hidden` to keep it out of GUIs.

An example can be found in the unit test for this feature.